Create step1_reading_akari_PCXV.pro
This is the routine for loading the data into IDL.

It imports HEALPix .FITS image files into arrays. Currently all the data from all the wavelengths and instruments is in HEALPIX, except for the AKARI/IRC 9 um data.

This step also applies basic image processing to the arrays: background subtraction and smoothing, for example.