Create step2_degrading_akari_PCXV.pro
This routine degrades the resolution of the iamges loaded during step1. More recently, since I implemented a simple gaussian pixel smoothing step in step1, step2 has become unnecessary.  We will need step2 in the futhre though, if we again compare images from instruments with similar resolution (such that the shape of the point-spread function becomes important). Now, everything is smoothed with a large guassian beam in step1, so we don't care so much about the convolution kernels between instruments.