step4_extractingSED_akari_PCXV
This is the step where the data gets arranged into SEDs- basically it stacks the images of all the wavelengths into a cube. One SED per pixel. There's a rescaling step too in case you don't need all of the individual pixels, and want to sample the SEDs at a larger scale.

I experimented before with combing this code with DustEM but ran into other issues.