Update step1_reading_akari_PCXV.py
Removed several steps from the original IDL process: The code for determining the size of each "cutout" and displaying it. The logic for doing this is that we are now doing the analysis based on the circular aperture photometry on the all-sky healpix images. So there's no need to make the cut-outs and display them individually.

However we might need to revisit the cut-out step if we move to higher resolution maps. If we do so, then we might run out of memory (and time..) trying to load all the HEALPix images into RAM and operate on them directly. Instead we may first make the cut outs of all the regions, and then loop through the cutouts one by one (like we did before).

ANother reason we might have to add this "cutouts" code is that me may or may not be able to access HEALPix maps for all of the data we want to analyze. The critical point here is the IRC 9um map which is still being finalized.