Create stepR_regridding.pro
This is just a simple routine to regrid and align images. 
The purpose is to put the AKRI/IRC (9 and 18 um) images on the same grid as the SPIRE 250 um image. 
The example here is rho Ophiuchi, until I find Herschel data for other regions as well.
I am writing it first in IDL, and then translating it to Python afterwards.