Update step0_settings.py
I found that "healpy" (the python implementation of HEALPix) automatically converts healpix maps from NESTED to RING.
I'm not sure that this is necessary, and it takes a long time. 
Now setting the "NESTED = TRUE" keyword when reading the healpix maps, so that this doesn't happen.
If I need the RING format later for some reason, I'll just remove this keyword and then make hard-copies of the fits files, in the RING scheme.