Update step1_reading_akari_PCXV.py
Adding the "nest" keyword, according to the commit made to the step0_settings.py (See the comment copied below from that commit).

=========7/17/15@11AM: step0_settings.py ==========================
I found that "healpy" (the python implementation of HEALPix) automatically converts healpix maps from NESTED to RING.
I'm not sure that this is necessary, and it takes a long time. 
Now setting the "nest=true" keyword when reading the healpix maps, so that this doesn't happen.
If I need the RING format later for some reason, I'll just remove this keyword and then make hard-copies of the fits files, in the RING scheme.