Update step1_reading_akari_PCXV.py
I found that hp.query_disc can't accept lists as inputs (like numpy functions). I guess this is a common issue with non-numpy functions?

A for loop seemed necessary, so I defined empty arrays for the three radii we want to know the pixels of. Then I did a loop over all of the ROIs, ".append"ing the result of each ROI to an empty array. The result is three lists (one for each radii), where the elements of the lists are lists of the desired pixel numbers.