Create stepH_haper.pro
This IDL procedure is for testing Clive Dickinson et al (2011)'s HEALPix photometry code on the AKARI maps.

The photometry routines themselves are left unchanged. This script just feeds the AKARI and IRAS data into them, to see that they work exactly as written.

This is a kind of short-cut accomplishing our current analysis goals (since the translation to Python is very time-consuming, and not likely to be finished before we have a chance to discuss the project at the IAU).