Create stepMH_ModifyHeader.py
This is a simple code for adding/modifying a header keyword using astropy.fits.

It's relevant to the stepH_haper.pro file, so I decided to copy it into this repository for reference.

Basically, it fixes a problem where the AKARI and IRAS HEALPix data don't have the proper keywords needed for them to be processed by Clive Dickinson's Healpix Photometry code, "multiepoch_photometry.pro". This IDL code was written for processing Planck maps, which are described in frequency terms. AKARI and IRAS use wavelength.