Update multiepoch_photometry.pro
I added the AKARI 65 um band's corresponding frequency, 4612 GHz, to the array of "allowed" frequencies, at the top of the code.
I also added a corresponding FWHM of the FIS PSF at 65 um. It's about 30 arcseconds, but I'll confirm the exact value later.
After testing at 65 um, I will also add in the other AKARI bands' values, and those of the IRAS bands, so that everything can be processed together.

At first glance, what I changed here is all that needs to be changed. Looking forward to whatever challenge that comes next, as a result of such a silly simple first assumption!