Update step5_BBfit.pro
Added comment at the ModBB definition about units Jy vs. MJy/sr

Generally speaking: Making big changed as to how the data is fed. Before, this routine was responsible for directly analyzing the images accoss all relevant wavelengths. Now, HAPER has done this already, and has made us a nice new .sav file that contains the photometry results. Yes, the circular aperture results are probably a bit more crude than the original step5 code. That is however the point. We want to check if our results are the same when we use the more robust circular apertures with annulus background subtraction and noise estimation, as well as the 1 degree smoothing.

This means: everything is much simpler. For example, in section 2 we defined arrays to hold the results for each pixel. Now, we just have one value per region. This also means that there is no "average values calculation section".

Except for the pixel-by-pixel to aperture changes however, most of the calculations will be the same.

For the arrays that we do need, the ones that store the BB fitting results of all the regions, I changed their dimensions to be determined from the number of maps (nmaps) and the number of regions (ns) which were supplied to HAPER.