Create stepSH_smooth_healpix.py
Created this new routine. The only purpose of this code is to do a one-tie smoothing of the Healpix maps needed for the AME:IR comparison. It makes 1-deg gaussian-smoothed versions of the AKARI FIS, IRAS, and Planck HFI Healpix maps. The purpose of using such a wide smoothing is to match the course FWHM of the AME componenet map from the Planck 2015 data release.